requirements.txt
# Pineapple Pipeline v1.0.0 -- tool dependencies
# Install: pip install -r requirements.txt

# Core (required for all tools)
pydantic>=2.0,<3.0
pyyaml>=6.0,<7.0

# pineapple doctor (health checks)
httpx>=0.27,<1.0

# Development (tests, linting)
pytest>=8.0,<9.0
ruff>=0.8,<1.0
